MODDCB-49 Add check for Request cancellation status
