+++ src/test/java/org/folio/dcb/service/BorrowingLibraryServiceTest.java
@@ -69,16 +69,6 @@ class BorrowingLibraryServiceTest {
   }
 
   @Test
-  void testTransactionCancelTest(){
-    var transactionEntity = createTransactionEntity();
-    transactionEntity.setStatus(OPEN);
-    TransactionStatus transactionStatus = TransactionStatus.builder().status(CANCELLED).build();
-    borrowingLibraryService.updateTransactionStatus(transactionEntity, transactionStatus);
-    verify(circulationService).cancelRequestIfExistOrNull(any());
-    Assertions.assertEquals(CANCELLED, transactionEntity.getStatus());
-  }
-
-  @Test
   void createTransactionTest() {
     borrowingLibraryService.createCirculation(DCB_TRANSACTION_ID, createDcbTransactionByRole(BORROWER), PICKUP_SERVICE_POINT_ID);
